requirements.txt
coverage
numpy==1.23.1
pandas==1.1.3
scikit-learn==1.1.1
scipy==1.10.0
mysql-connector-python==8.0.33